refactor: change artist entity name
