+++ src/artist/artist.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ArtistController } from './artist.controller';
 import { ArtistService } from './artist.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ArtistEntity } from 'src/core/entitys/main/artistEntity';
+import { ArtistEntity } from 'src/core/entitys/main/artist.entity';
 import { SongEntity } from 'src/core/entitys/main/song.entity';
 
 @Module({
